+++ lib/active_interaction/base.rb
@@ -31,7 +31,7 @@ module ActiveInteraction
     include ActiveModelable
     include Runnable
 
-    validate :input_errors
+    define_callbacks :type_check
 
     class << self
       include Hashable
@@ -212,6 +212,14 @@ module ActiveInteraction
       end
     end
 
+    protected
+
+    def run_validations!
+      type_check
+
+      super if errors.empty?
+    end
+
     private
 
     # @param inputs [Hash{Symbol => Object}]
@@ -234,16 +242,16 @@ module ActiveInteraction
         begin
           public_send("#{name}=", filter.clean(inputs[name]))
         rescue InvalidValueError, MissingValueError, InvalidNestedValueError
-          # Validators (#input_errors) will add errors if appropriate.
+          # #type_check will add errors if appropriate.
         end
       end
     end
 
-    # @!group Validations
-
-    def input_errors
-      Validation.validate(self.class.filters, inputs).each do |error|
-        errors.add_sym(*error)
+    def type_check
+      run_callbacks(:type_check) do
+        Validation.validate(self.class.filters, inputs).each do |error|
+          errors.add_sym(*error)
+        end
       end
     end
   end